round 3? of drf
Each node will build 1 tree at a time (concurrent with other nodes), and
will be the master for that tree, pulling from an atomic tree list as it
finishes its current tree.  Each tree is built on all nodes, but the
master for a tree varies.  Each node maintains a local list of rows to
be used for each tree; the local list is ordered in sections by tree
split.  The tree-split bounds are kept on the master.  Local node work
is limited to computing histograms over ranges of rows, or sorting a
split into low & high halves.