Fix bug with sending _drf._kids fields & resends
After the initial send-attempt, the nested kid fields begin to get
expanded.  However if the network fails, on a resend - the same kids
(now expanded) are sent again - but with a different size.  Blows the
resend invariants.  Fixed by just sending the interesting fields instead
of the whole kids fields.